Cleanup common Read code in HttpRequestStream
