+++ src/System.Net.HttpListener/src/System/Net/Managed/HttpRequestStream.Managed.cs
@@ -64,18 +64,6 @@ namespace System.Net
         // -1 if we had a content length set and we finished reading that many bytes.
         private int FillFromBuffer(byte[] buffer, int offset, int count)
         {
-            if (buffer == null)
-                throw new ArgumentNullException(nameof(buffer));
-            if (offset < 0)
-                throw new ArgumentOutOfRangeException(nameof(offset));
-            if (count < 0)
-                throw new ArgumentOutOfRangeException(nameof(count));
-            int len = buffer.Length;
-            if (offset > len)
-                throw new ArgumentException(nameof(offset), SR.offset_out_of_range);
-            if (offset > len - count)
-                throw new ArgumentException(nameof(count), SR.offset_out_of_range);
-
             if (_remainingBody == 0)
                 return -1;
 
@@ -101,23 +89,10 @@ namespace System.Net
             return size;
         }
 
-        public override int Read(byte[] buffer, int offset, int size)
+        protected virtual int ReadCore(byte[] buffer, int offset, int size)
         {
             if (_disposed)
                 throw new ObjectDisposedException(typeof(HttpRequestStream).ToString());
-            if (buffer == null)
-            {
-                throw new ArgumentNullException(nameof(buffer));
-            }
-            if (offset < 0 || offset > buffer.Length)
-            {
-                throw new ArgumentOutOfRangeException(nameof(offset));
-            }
-            if (size < 0 || size > buffer.Length - offset)
-            {
-                throw new ArgumentOutOfRangeException(nameof(size));
-            }
-
             // Call FillFromBuffer to check for buffer boundaries even when remaining_body is 0
             int nread = FillFromBuffer(buffer, offset, size);
             if (nread == -1)
@@ -135,21 +110,11 @@ namespace System.Net
             return nread;
         }
 
-        public override IAsyncResult BeginRead(byte[] buffer, int offset, int size, AsyncCallback cback, object state)
+        protected virtual IAsyncResult BeginReadCore(byte[] buffer, int offset, int size, AsyncCallback cback, object state)
         {
             if (_disposed)
                 throw new ObjectDisposedException(typeof(HttpRequestStream).ToString());
-            if (buffer == null)
             {
-                throw new ArgumentNullException(nameof(buffer));
-            }
-            if (offset < 0 || offset > buffer.Length)
-            {
-                throw new ArgumentOutOfRangeException(nameof(offset));
-            }
-            if (size < 0 || size > buffer.Length - offset)
-            {
-                throw new ArgumentOutOfRangeException(nameof(size));
             }
 
             int nread = FillFromBuffer(buffer, offset, size);
